Fix feature compilation for `consensus`
- Fix feature compilation for `consensus`.
- Move `ChunksPushResult` and `ChunksPushError` from the `blockchain`
  crate to the `blockchain-interface` crate such that they can be
  used by other crates (specially `consensus`) without needing to
  pull `blockchain` crate dependencies.
