+++ blockchain-interface/src/lib.rs
@@ -1,7 +1,10 @@
 pub use abstract_blockchain::AbstractBlockchain;
 pub use chain_info::ChainInfo;
 pub use chain_ordering::*;
-pub use error::{BlockchainError, BlockchainEvent, Direction, ForkEvent, PushError, PushResult};
+pub use error::{
+    BlockchainError, BlockchainEvent, ChunksPushError, ChunksPushResult, Direction, ForkEvent,
+    PushError, PushResult,
+};
 
 mod abstract_blockchain;
 mod chain_info;
