+++ blockchain/src/blockchain/mod.rs
@@ -2,7 +2,6 @@ mod abstract_blockchain;
 pub mod accounts;
 #[allow(clippy::module_inception)]
 pub mod blockchain;
-pub mod error;
 pub mod history_sync;
 pub mod inherents;
 pub mod push;
