+++ blockchain/src/chain_metrics.rs
@@ -1,13 +1,11 @@
 use nimiq_block::Block;
 use nimiq_block::BlockBody::Micro;
-use nimiq_blockchain_interface::{PushError, PushResult};
+use nimiq_blockchain_interface::{ChunksPushError, ChunksPushResult, PushError, PushResult};
 use nimiq_hash::Blake2bHash;
 use prometheus_client::encoding::text::Encode;
 use prometheus_client::metrics::counter::Counter;
 use prometheus_client::metrics::family::Family;
 use prometheus_client::registry::Registry;
-
-use crate::blockchain::error::{ChunksPushError, ChunksPushResult};
 
 #[derive(Default)]
 pub struct BlockchainMetrics {
