+++ blockchain/src/lib.rs
@@ -2,7 +2,6 @@
 extern crate log;
 
 pub use blockchain::blockchain::{Blockchain, BlockchainConfig, TransactionVerificationCache};
-pub use blockchain::error::*;
 pub use history::*;
 
 pub(crate) mod blockchain;
